Add new test suites for graph states
