+++ examples/preact-graph-state/.gitignore
@@ -1,0 +1,4 @@
+node_modules
+.DS_Store
+dist
+*.local
